Move ResourceTable/NativeResource to Graphics lib
